google map pour chaque team avec un marqueur clicable et fonctionnelle

+++ db/migrate/20201204105403_create_compets.rb
@@ -2,8 +2,8 @@ class CreateCompets < ActiveRecord::Migration[5.2]
   def change
     create_table :compets do |t|
       t.string :address
-      t.float :latitude
-      t.float :longitude
+      t.decimal :latitude
+      t.decimal :longitude
 
       t.timestamps
     end
